feat: Navigation implementation
