+++ composeApp/src/commonMain/kotlin/com/example/mokumokusolo/ui/main/MainScreen.kt
@@ -2,6 +2,9 @@ package com.example.mokumokusolo.ui.main
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
@@ -10,6 +13,9 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavDestination.Companion.hierarchy
@@ -18,8 +24,11 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
+import com.example.mokumokusolo.model.App
+import com.example.mokumokusolo.model.Expenditure
 import com.example.mokumokusolo.navigation.AppDestination
 import com.example.mokumokusolo.navigation.bottomNavItems
+import com.example.mokumokusolo.ui.screen.addItem.AddItemScreen
 import com.example.mokumokusolo.ui.screen.home.HomeScreen
 
 @Composable
@@ -28,6 +37,9 @@ fun MainScreen(modifier: Modifier = Modifier) {
     val currentBackStack by navController.currentBackStackEntryAsState()
     val currentDestination = currentBackStack?.destination
 
+    var showAddItemScreen by remember { mutableStateOf(false) }
+    var apps by remember { mutableStateOf(emptyList<App>()) }
+    var expenditures by remember { mutableStateOf(emptyList<Expenditure>()) }
 
     Scaffold(
         modifier = modifier
@@ -69,6 +81,13 @@ fun MainScreen(modifier: Modifier = Modifier) {
                     )
                 }
             }
+        },
+        floatingActionButton = {
+            FloatingActionButton(
+                onClick = { showAddItemScreen = true }
+            ) {
+                Icon(Icons.Default.Add, "")
+            }
         }
     ) {
         NavHost(
@@ -77,11 +96,34 @@ fun MainScreen(modifier: Modifier = Modifier) {
         ) {
             composable<AppDestination.Home> {
                 HomeScreen(
+                    apps = apps,
+                    expenditures = expenditures,
                 )
             }
             composable<AppDestination.Calendar> {
 //                CalendarScreen()
             }
+        }
+
+        if (showAddItemScreen) {
+            AddItemScreen(
+                onClose = { showAddItemScreen = false },
+                onAddItem = { isIncome, name, amount ->
+                    if (isIncome) {
+                        val newApp =
+                            App(id = apps.size + 1, name = name, amount = amount.toDouble())
+                        apps = apps + newApp
+                    } else {
+                        val newExpenditure = Expenditure(
+                            id = expenditures.size + 1,
+                            name = name,
+                            amount = amount.toDouble()
+                        )
+                        expenditures = expenditures + newExpenditure
+                    }
+                    showAddItemScreen = false
+                }
+            )
         }
     }
 }
