+++ composeApp/src/commonMain/kotlin/com/example/mokumokusolo/ui/screen/home/HomeScreen.kt
@@ -1,19 +1,11 @@
 package com.example.mokumokusolo.ui.screen.home
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -24,104 +16,78 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.example.mokumokusolo.model.App
 import com.example.mokumokusolo.model.Expenditure
-import com.example.mokumokusolo.ui.screen.addItem.AddItemScreen
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
-fun HomeScreen(modifier: Modifier = Modifier) {
+fun HomeScreen(
+    apps: List<App> = emptyList(),
+    expenditures: List<Expenditure> = emptyList(),
+    modifier: Modifier = Modifier
+) {
     var selectedChip by remember { mutableStateOf("all") }
-    var showAddItemScreen by remember { mutableStateOf(false) }
 
-    var apps by remember { mutableStateOf(emptyList<App>()) }
-    var expenditures by remember { mutableStateOf(emptyList<Expenditure>()) }
-
-    if (showAddItemScreen) {
-        AddItemScreen(
-            onClose = { showAddItemScreen = false },
-            onAddItem = { isIncome, name, amount ->
-                if (isIncome) {
-                    val newApp = App(id = apps.size + 1, name = name, amount = amount.toDouble())
-                    apps = apps + newApp
-                } else {
-                    val newExpenditure = Expenditure(
-                        id = expenditures.size + 1,
-                        name = name,
-                        amount = amount.toDouble()
-                    )
-                    expenditures = expenditures + newExpenditure
-                }
-                showAddItemScreen = false
-            }
+    Column(
+        modifier = modifier
+            .padding(top = 16.dp)
+            .fillMaxWidth(),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        HomeCards(
+            apps = apps,
+            expenditures = expenditures
         )
-    } else {
-        Scaffold(
-            modifier = modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.background),
-            floatingActionButton = {
-                FloatingActionButton(
-                    onClick = { showAddItemScreen = true },
-                    containerColor = MaterialTheme.colorScheme.primaryContainer
-                ) {
-                    Icon(Icons.Default.Add, "")
-                }
-            }
-        ) { paddingValues ->
-            Column(
-                modifier = Modifier
-                    .padding(paddingValues)
-                    .padding(top = 16.dp)
-                    .fillMaxWidth(),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                HomeCards(
-                    apps = apps,
-                    expenditures = expenditures
+        Row(
+            horizontalArrangement = Arrangement.SpaceAround,
+            modifier = Modifier.padding(top = 16.dp)
+        ) {
+            AllChip(
+                selected = selectedChip == "all",
+                onClicked = { selectedChip = "all" }
+            )
+            AppsChip(
+                selected = selectedChip == "apps",
+                onClicked = { selectedChip = "apps" }
+            )
+            ExpenditureChip(
+                selected = selectedChip == "expenditure",
+                onClicked = { selectedChip = "expenditure" }
+            )
+        }
+        when (selectedChip) {
+            "all" -> {
+                AppList(
+                    appList = apps,
+                    modifier = Modifier.width(300.dp)
                 )
-                Row(
-                    horizontalArrangement = Arrangement.SpaceAround,
-                    modifier = Modifier.padding(top = 16.dp)
-                ) {
-                    AllChip(
-                        selected = selectedChip == "all",
-                        onClicked = { selectedChip = "all" }
-                    )
-                    AppsChip(
-                        selected = selectedChip == "apps",
-                        onClicked = { selectedChip = "apps" }
-                    )
-                    ExpenditureChip(
-                        selected = selectedChip == "expenditure",
-                        onClicked = { selectedChip = "expenditure" }
-                    )
-                }
-                when (selectedChip) {
-                    "all" -> {
-                        AppList(
-                            appList = apps,
-                            modifier = Modifier.width(300.dp)
-                        )
-                        ExpenditureList(
-                            expenditureList = expenditures,
-                        )
-                    }
-
-                    "apps" -> AppList(
-                        appList = apps,
-                    )
-
-                    "expenditure" -> ExpenditureList(
-                        expenditureList = expenditures,
-                    )
-                }
+                ExpenditureList(
+                    expenditureList = expenditures,
+                )
             }
+
+            "apps" -> AppList(
+                appList = apps,
+            )
+
+            "expenditure" -> ExpenditureList(
+                expenditureList = expenditures,
+            )
         }
     }
 }
 
+
 @Preview
 @Composable
 fun HomeScreenPreview() {
-    HomeScreen()
+    HomeScreen(
+        apps = listOf(
+            App(1, "MokuMokuSolo", 1000.0),
+            App(2, "AnotherApp", 2000.0)
+        ),
+        expenditures = listOf(
+            Expenditure(1, "Netflix", 500.0),
+            Expenditure(2, "Spotify", 300.0)
+        )
+    )
 }
